backend/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Manuel Gil <manuelgilsitio@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlmodel = "^0.0.22"
fastapi = {extras = ["standard"], version = "^0.115.5"}
passlib = "^1.7.4"
bcrypt = "^4.2.1"
pyjwt = "^2.10.0"
itsdangerous = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
